+++ outfile/include/epanet_output_enums.h
