+++ shared/CMakeLists.txt
@@ -1,5 +1,5 @@
 #
-# CMakeLists.txt - CMake configuration file for targets/util
+# CMakeLists.txt - CMake configuration file for shared
 #
 # Created: July 19, 2019
 #
@@ -9,23 +9,23 @@
 
 
 
-set(UTIL_SOURCES
+set(SHARED_SOURCES
     cstr_helper.c
     errormanager.c
     filemanager.c
     list.c
     )
 
-set(UTIL_HEADERS
+set(SHARED_HEADERS
     cstr_helper.h
     errormanager.h
     filemanager.h
     list.h
     )
 
-add_library(util_objs OBJECT ${UTIL_SOURCES})
+add_library(shared_objs OBJECT ${SHARED_SOURCES})
 
-target_include_directories(util_objs
+target_include_directories(shared_objs
     PUBLIC
         ${CMAKE_CURRENT_SOURCE_DIR}
     )
