+++ shared/cstr_helper.c
