+++ shared/cstr_helper.h
