+++ shared/errormanager.c
