+++ shared/errormanager.h
