+++ shared/filemanager.c
