+++ shared/filemanager.h
