+++ shared/list.c
