+++ solver/enumstxt.h
