+++ solver/funcs.h
