+++ solver/genmmd.c
