+++ solver/hydcoeffs.c
