+++ solver/inpfile.c
