+++ solver/input1.c
