+++ solver/input2.c
