+++ solver/input3.c
