+++ solver/output.c
