+++ solver/quality.c
