+++ solver/qualreact.c
