+++ solver/rules.c
