+++ solver/smatrix.c
