+++ solver/text.h
