+++ solver/types.h
